Move NER to ingest-file service
