+++ Dockerfile
@@ -23,22 +23,6 @@ RUN groupadd -g 1000 -r app \
 # Install Python dependencies
 RUN pip3 install --no-cache-dir -q -U pip setuptools six
 
-# Install spaCy and link models to three-letter language codes
-RUN pip3 install --no-cache-dir -q spacy==2.2.3
-RUN python3 -m spacy download xx_ent_wiki_sm \
-    && python3 -m spacy link xx_ent_wiki_sm xx
-RUN python3 -m spacy download en_core_web_sm \
-    && python3 -m spacy link en_core_web_sm eng
-RUN python3 -m spacy download de_core_news_sm \
-    && python3 -m spacy link de_core_news_sm deu
-RUN python3 -m spacy download fr_core_news_sm \
-    && python3 -m spacy link fr_core_news_sm fra
-RUN python3 -m spacy download es_core_news_sm \
-    && python3 -m spacy link es_core_news_sm spa
-RUN python3 -m spacy download pt_core_news_sm \
-    && python3 -m spacy link pt_core_news_sm por
-ENV ALEPH_NER_MODELS=eng:deu:fra:spa:por
-
 COPY requirements-generic.txt /tmp/
 RUN pip3 install --no-cache-dir -r /tmp/requirements-generic.txt
 COPY requirements-toolkit.txt /tmp/
@@ -54,7 +38,6 @@ RUN pip install -e /aleph
 # Configure some docker defaults:
 ENV ALEPH_ELASTICSEARCH_URI=http://elasticsearch:9200/ \
     ALEPH_DATABASE_URI=postgresql://aleph:aleph@postgres/aleph \
-    ALEPH_GEONAMES_DATA=/aleph/contrib/geonames.txt \
     BALKHASH_BACKEND=postgresql \
     BALKHASH_DATABASE_URI=postgresql://aleph:aleph@postgres/aleph \
     REDIS_URL=redis://redis:6379/0 \
