+++ aleph/settings.py
@@ -92,19 +92,6 @@ UI_LANGUAGES = ['ru', 'es', 'de', 'bs', 'en']
 UI_LANGUAGES = env.to_list('ALEPH_UI_LANGUAGES', UI_LANGUAGES)
 UI_LANGUAGES = [l.lower().strip() for l in UI_LANGUAGES]
 
-# Geonames data file
-GEONAMES_DATA = env.get('ALEPH_GEONAMES_DATA')
-
-# FastText lid model file
-LID_MODEL_PATH = env.get('ALEPH_LID_MODEL_PATH', '/aleph/contrib/lid.176.ftz')
-
-# Disable entity extraction
-TAG_ENTITIES = env.to_bool('ALEPH_TAG_ENTITIES', True)
-
-# List available NER models
-NER_MODELS = set(env.to_list('ALEPH_NER_MODELS', ['eng']))
-NER_DEFAULT_MODEL = 'xx'
-
 # Result high-lighting
 RESULT_HIGHLIGHT = env.to_bool('ALEPH_RESULT_HIGHLIGHT', True)
