+++ docker-compose.dev.yml
@@ -39,7 +39,7 @@ services:
     restart: on-failure
     tmpfs:
       - /tmp
-      - /root/.config
+      - /home/app/.config
 
   ingest-file:
     build:
@@ -53,6 +53,7 @@ services:
       - archive-data:/data
       - "./services/ingest-file/ingestors:/ingestors/ingestors"
       - "./services/ingest-file/tests:/ingestors/tests"
+      - "./services/ingest-file/data:/ingestors/data"
       - "./services/ingest-file/requirements.txt:/ingestors/requirements.txt"
       - "/:/host"
     links:
