+++ requirements-generic.txt
@@ -27,7 +27,6 @@ redis==3.3.11
 fakeredis==1.1.0
 networkx==2.4
 zipstream==1.1.4
-fasttext==0.9.1
 tabulate==0.8.6
 
 grpcio==1.26.0
