+++ services/ingest-file/Dockerfile
@@ -114,6 +114,21 @@ RUN pip3 install --no-cache-dir -q -U pip setuptools six wheel nose coverage
 COPY requirements.txt /tmp/
 RUN pip3 install --no-cache-dir -r /tmp/requirements.txt
 
+# Install spaCy and link models to three-letter language codes
+RUN python3 -m spacy download xx_ent_wiki_sm \
+    && python3 -m spacy link xx_ent_wiki_sm xx
+RUN python3 -m spacy download en_core_web_sm \
+    && python3 -m spacy link en_core_web_sm eng
+RUN python3 -m spacy download de_core_news_sm \
+    && python3 -m spacy link de_core_news_sm deu
+RUN python3 -m spacy download fr_core_news_sm \
+    && python3 -m spacy link fr_core_news_sm fra
+RUN python3 -m spacy download es_core_news_sm \
+    && python3 -m spacy link es_core_news_sm spa
+RUN python3 -m spacy download pt_core_news_sm \
+    && python3 -m spacy link pt_core_news_sm por
+ENV INGESTORS_NER_MODELS=eng:deu:fra:spa:por
+
 COPY . /ingestors
 WORKDIR /ingestors
 RUN pip3 install --no-cache-dir -e /ingestors
@@ -124,7 +139,7 @@ ENV ARCHIVE_TYPE=file \
     BALKHASH_BACKEND=postgresql \
     BALKHASH_DATABASE_URI=postgresql://aleph:aleph@postgres/aleph \
     REDIS_URL=redis://redis:6379/0 \
-    UNOSERVICE_URL=http://convert-document:3000/convert
+    INGESTORS_CONVERT_DOCUMENT_URL=http://convert-document:3000/convert
 
 # USER app
 CMD ingestors process
