+++ services/ingest-file/requirements.txt
@@ -13,6 +13,9 @@ google-cloud-vision==0.41.0
 google-cloud-storage==1.24.1
 
 tesserocr==2.5.0
+spacy==2.2.3
+fingerprints==0.6.6
+fasttext==0.9.1
 
 # Development
 nose==1.3.7
